requirements.txt
discord
pyowm
python-dotenv
